Add diffing for users, updates and renames

+++ src/Accord.Services/UserMessages/DeleteMessage.cs
@@ -19,7 +19,11 @@ namespace Accord.Services.UserMessages
 
         protected override async Task Handle(DeleteMessageRequest request, CancellationToken cancellationToken)
         {
-            var message = await _db.UserMessages.SingleAsync(x => x.Id == request.DiscordMessageId, cancellationToken: cancellationToken);
+            var message = await _db.UserMessages.SingleOrDefaultAsync(x => x.Id == request.DiscordMessageId, cancellationToken: cancellationToken);
+
+            if (message is null)
+                return;
+
             _db.Remove(message);
             await _db.SaveChangesAsync(cancellationToken);
         }
